+++ lib/data_seeder/loader/csv.rb
@@ -11,8 +11,15 @@ module DataSeeder
         @line_number = 0
         csv = ::CSV.new(io, headers: true)
         csv.each do |row|
-          @line_number += 1
-          save(row.to_hash)
+          begin
+            @line_number += 1
+            save(row.to_hash)
+          rescue Exception => e
+            # TODO: Consider counting the header in the line_number count, but anyone using
+            # config[:use_line_number_as_id] would have all there id's incremented
+            logger.error "Exception at line #{@line_number+1}: #{e.message}"
+            raise unless config[:continue_on_exception]
+          end
         end
       end
     end