+++ History.md
@@ -1,9 +1,15 @@
 data_seeder Changelog
 =====================
 
+1.0.3 / 2016-02-19
+  - Log messages are now info instead of debug.
+  - DataSeeder.test_run deprecated in favor of quiet_run which logs to Rails.log instead of $stdout
+  - Config continue_on_exception added for txt,csv.
+  - No longer use $LINE_NUMBER to count line numbers as JRuby has issues.
+  
 1.0.2 / 2016-02-17
   - Bug fix the default_config stuff.
-  
+
 1.0.1 / 2016-02-17
   - Deprecate class method default_config and prefer instance method instead
     so it can be used in manual seeding.